+++ jedani-worker/src/main/java/ru/complitex/jedani/worker/service/RewardService.java
@@ -639,7 +639,11 @@ public class RewardService implements Serializable {
             updateLocal(sale, reward, period);
 
             if (reward.getPoint().compareTo(ZERO) != 0) {
-                domainService.save(reward);
+                if (test) {
+                    rewards.add(reward);
+                } else {
+                    domainService.save(reward);
+                }
             }
         }
 
@@ -1190,11 +1194,11 @@ public class RewardService implements Serializable {
     }
 
     @Transactional(rollbackFor = RewardException.class)
-    public void calculateRewards() throws RewardException {
+    public void calculateRewards(Period period) throws RewardException {
         try {
-            Period period = periodMapper.getActualPeriod();
-
-            rewardMapper.deleteRewards(period.getObjectId());
+            if (!test) {
+                rewardMapper.deleteRewards(period.getObjectId());
+            }
 
             rewards.clear();
 
@@ -1210,6 +1214,7 @@ public class RewardService implements Serializable {
 
             saleService.getSales().stream()
                     .filter(sale -> !isPaidSalePeriod(sale))
+                    .filter(sale -> sale.getPeriodId() <= period.getObjectId())
                     .forEach(sale -> {
                         calculateSaleReward(sale, saleService.getSaleItems(sale.getObjectId()), period, RewardStatus.ESTIMATED);
                         calculateSaleReward(sale, saleService.getSaleItems(sale.getObjectId()), period, RewardStatus.CHARGED);
@@ -1279,10 +1284,15 @@ public class RewardService implements Serializable {
         }
     }
 
-    public void testRewards() throws RewardException {
+    @Transactional(rollbackFor = RewardException.class)
+    public void calculateRewards() throws RewardException {
+        calculateRewards(periodMapper.getActualPeriod());
+    }
+
+    public void testRewards(Period period) throws RewardException {
         test = true;
 
-        calculateRewards();
+        calculateRewards(period);
 
         test = false;
     }
